Master Data Frame (High Resolution)

A mapped spectral data frame created using createSpecDF() and mapSpectrum(). Contains the baseline corrected (by linear interpolation) intensity data for 6 spectra, with m/z data recorded up to 4 decimal places.

data("Master")

Format

A data frame with 980506 observations on the following 7 variables.

full_mz

A numeric vector; the m/z data for all spectra in the data frame.

Blank1

A numeric vector; the intensity data for the first spectrum of the dataset. A MALDI matrix blank spectrum.

Blank2

A numeric vector; the intensity data for the second spectrum in the dataset. A MALDI matrix blank spectrum.

Before1

A numeric vector; the intensity data for the third spectrum in the dataset. Spectrum obtained prior to chemical intervention.

Before2

A numeric vector; the intensity data for the fourth spectrum in the dataset. Spectrum obtained prior to chemical intervention.

After1

A numeric vector; the intensity data for the fifth spectrum in the dataset. Spectrum obtained after chemical intervention.

After2

A numeric vector; the intensity data for the sixth spectrum in the dataset. Spectrum obtained after chemical intervention.

Source

Yeh, K., Stock N. L., Burr, W. & Stotesbury, T. Preliminary analysis of latent fingerprints recovered from underneath bloodstains using Matrix-Assisted Laser Desportion/Ionization Fourier-Transform Ion Cyclotron Resonance Mass Spectrometry Imaging (MALDI FT-ICR MSI). Forensic Chemistry (2020). In press.

References

https://github.com/wesleyburr/subMaldi

Examples

data(Master)

## Separate samples
blanks <- dplyr::select(Master, "full_mz", "Blank1", "Blank2")
precon <- dplyr::select(Master, "full_mz", "Before1", "Before2")
postcon <- dplyr::select(Master, "full_mz", "After1", "After2")