Find Most Intense Peak of a Set

Analyzes spectral data and returns information about the most intense peak in all of the spectral set. Indicates the most intense spectrum, the m/z value, and intesnity of the maximum. User must specify at least two spectra.

find_max_set(dat, mass_dat, spectra_cols)

Arguments

dat
The name of the spectral data frame, containing m/z data in the first column and spectral intensity data in subsequent columns.

mass_dat
A character string; the name of the column in dat containing the m/z data for the spectrum.

spectra_cols
A character vector; the names of the column in dat containing the intensity data for the spectra-of-interest.

Value

Returns a data frame indidcating the peak maximum of the spectral set. The sample number, m/z value, and intensity data are returned.

References

https://github.com/wesleyburr/subMaldi

Author

Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wesleyburr@trentu.ca> Sophie Castel <sophie.castel@ontariotechu.net>

Examples

## Load sample dataset "Master.rda"
data("Master")

## Find the most intense peak of six spectra
find_max_set(dat = Master, mass_dat = "full_mz", spectra_cols = c("After1", "Blank2", 
            "Before1", "Blank1", "After2", "Before2"))

#>        Intensity (Max)     Mass
#> Blank1     58818870554 248.9629