Plot a Single Spectrum

Plots a single spectrum from a mapped spectral data frame.

plotSpectrum(dat, mass_dat, intensity_dat, colour = "slategrey", 
            span = 5, thresh = 0.1, min_mz = 0, max_mz = 1000, 
            lbls = FALSE, lbl.fmt = "%3.4f",x_ticks = 100)

Arguments

dat
The mapped spectral data frame, containing full_mz in the first column.

mass_dat
The name of the column in dat containing the m/zdata for the spectrum that is to be plotted. Should be input as dat$mass_column. By default, the mass column in a mapped data frame is "full_mz". If our spectral data frame is called my_spec, then our input for mass_dat = my_spec$full_mz.

intensity_dat
The name of the column in dat containing the intensity data for the spectrum that is to be plotted. Should be input as dat$intensity_column, as above for mass_dat.

colour
A character string; the name of the colour the spectrum should be plotted in. Default = "slategrey".

span
Single numeric value; the distance between peak maxima that are labelled if lbls = TRUE. Default = 5.

thresh
Single numeric value between 0 and 1; the intensity threshold of peaks in the plotted spectrum whih should be labelled. Default = 0.1.

min_mz
Single numeric value; minimum m/z value of the observed range. Default = 0.

max_mz
Single numeric value; upper end of m/z range to be observed in the plotted spectra. Default = 1000.

lbls
Logical; should labels indicating the m/z of the most intense peaks be included in the plot? Default = FALSE.

lbl.fmt
Character string in the format "%a.bf", where a is the number of figures displayed to the left of decimal places in labels, and b is the number of figures displayed to the right of decimal places in labels. Default = "%3.4f".

x_ticks
Single numeric value; the increment between ticks on the x-axis. Default = 100.

Details

Value

Returns a line plot of the input spectrum.

References

https://github.com/wesleyburr/subMaldi

Author

Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>

Note

See also

Examples

## Plotting using the sample dataset "Before1.rda"
data("Before1")

plotSpectrum(dat = Before1, mass_dat = Before1$mass, 
              intensity_dat = Before1$Intensity)

#> Error in plotSpectrum(dat = Before1, mass_dat = Before1$mass, intensity_dat = Before1$Intensity): could not find function "plotSpectrum"