Import .csv Spectra from a Directory and Export

Imports all .csv files in a directory, turns them into a binary data frame containing a "mass" and "Intensity" column, and outputs them as .rda files into an output directory. Allows for rapid import of many spectral datasets at once.

readcsvDir(direct, massCol, intenseCol, output)

Arguments

direct
The path to the directory where the .csv files are held.

massCol
A character string; the name of the mass column in the .csv files.

intenseCol
A character string; the name of the intensity column in the .csv files.

output
The path to the directory where the exported .rda files containing the binary spectral data frames should go.

Value

Returns a directory of .rda files containing binary spectral data frames into the output file path.

Details

Assumes that the .csv files are organized in tidy format, i.e., columns are variables and rows are individual m/z observations.

References

https://github.com/wesleyburr/subMaldi

Author

Kristen Yeh <kristenyeh@trentu.ca> Wesley Burr <wburr@trentu.ca>